Rename project, dirs for Python friendly import names
